# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TELEPIZZA",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "GET ALL CATEGORIES",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "1yTC5wISXg1IS43b7m2QtVqgJbp44ce4fnYdpCo9"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "General"
        ],
        "summary": "POST CREATE NEW CATEGORY",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "categoryName": "Salads",
                  "description": "Descubre nuestras nuevas ensaladas VIPS hechas con ingredientes tan sabrosos como aguacate, queso de cabra, tiras de vacuno marinado o mango. Waikiki, Santorini, Los \ufffdngeles o 5 Quesos ..."
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "1yTC5wISXg1IS43b7m2QtVqgJbp44ce4fnYdpCo9"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}